requirement.txt
flask
pyyaml
flask-sqlalchemy
flask-marshmallow
flask-script
flask-cors==3.0.2
gunicorn
ipython
pymysql
pyjwt
gevent
marshmallow-sqlalchemy
xlrd
xlwt
requests
oss2
httprunner==1.5.13
ordereddict
urllib3
flask-apscheduler
python-jenkins
flask-login
mysqlclient
pandas==0.24.2
jira
tornado
ujson
aredis
redis
flask-caching